# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TimeDive API — history learning through personalized AI stories
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me/profile:
    get:
      operationId: getProfile
      tags: [users]
      summary: Get current user's profile
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProfile
      tags: [users]
      summary: Update current user's profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/preferences:
    get:
      operationId: getPreferences
      tags: [users]
      summary: Get current user's interest preferences
      responses:
        "200":
          description: User preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
    put:
      operationId: updatePreferences
      tags: [users]
      summary: Update current user's interest preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PreferencesUpdate"
      responses:
        "200":
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UserProfile:
      type: object
      required: [id, email, displayName, onboardingComplete, createdAt]
      properties:
        id:
          type: number
        email:
          type: string
        displayName:
          type: string
        ageMode:
          type: ["string", "null"]
          enum: [kid, teen, adult, null]
        age:
          type: ["integer", "null"]
        gradeLevel:
          type: ["integer", "null"]
        onboardingComplete:
          type: boolean
        recapEmailOptIn:
          type: boolean
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: ["string", "null"]
          format: date-time
    ProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
        ageMode:
          type: string
          enum: [kid, teen, adult]
        age:
          type: ["integer", "null"]
        gradeLevel:
          type: ["integer", "null"]
        onboardingComplete:
          type: boolean
        recapEmailOptIn:
          type: boolean
    UserPreferences:
      type: object
      required: [userId]
      properties:
        userId:
          type: number
        sports:
          type: array
          items:
            type: string
        videoGames:
          type: array
          items:
            type: string
        movieGenres:
          type: array
          items:
            type: string
        hobbies:
          type: array
          items:
            type: string
        books:
          type: array
          items:
            type: string
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]
    PreferencesUpdate:
      type: object
      properties:
        sports:
          type: array
          items:
            type: string
        videoGames:
          type: array
          items:
            type: string
        movieGenres:
          type: array
          items:
            type: string
        hobbies:
          type: array
          items:
            type: string
        books:
          type: array
          items:
            type: string
        city:
          type: ["string", "null"]
        country:
          type: ["string", "null"]

    # Topic schemas
...